chore: settings

+++ apps/app-trade-agents/src/screens/Routes/RouteViewScreen.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useLayoutEffect, useRef } from 'react';
 import { RouteProp, useRoute, useScrollToTop } from '@react-navigation/native';
-import { View, FlatList } from 'react-native';
+import { View, FlatList, Alert } from 'react-native';
 import { Divider } from 'react-native-paper';
 import { useNavigation } from '@react-navigation/core';
 
@@ -65,7 +65,18 @@ const RouteViewScreen = () => {
         navigation.goBack();
       }
     };
-    deleteRoute();
+
+    Alert.alert('Вы уверены, что хотите удалить маршрут и его документы?', '', [
+      {
+        text: 'Да',
+        onPress: async () => {
+          deleteRoute();
+        },
+      },
+      {
+        text: 'Отмена',
+      },
+    ]);
   }, [dispatch, docDispatch, geoList, id, navigation, orderList, returnList, visitList]);
 
   const actionsMenu = useCallback(() => {
